# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant API built with hexagonal architecture. Provides identity management, tenant management, email templates, and more.",
        "title": "Hexagonal Go API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/oauth": {
            "get": {
                "description": "Redirects to the OAuth provider's authorization page. After authorization, the provider will redirect back to the callback endpoint.",
                "tags": [
                    "identity"
                ],
                "summary": "Initiate OAuth flow",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth provider name",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tenant ID for new user registration",
                        "name": "tenant_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Redirect to OAuth provider"
                    },
                    "400": {
                        "description": "Provider is required or unsupported",
                        "schema": {
                            "$ref": "#/definitions/internal_identity_interface_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to initiate OAuth",
                        "schema": {
                            "$ref": "#/definitions/internal_identity_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/oauth/callback": {
            "get": {
                "description": "Handles the callback from the OAuth provider after user authorization. Exchanges the authorization code for tokens and logs the user in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "identity"
                ],
                "summary": "OAuth callback",
                "parameters": [
                    {
                        "enum": [
                            "google",
                            "github"
                        ],
                        "type": "string",
                        "description": "OAuth provider name",
                        "name": "provider",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization code from OAuth provider",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "State token for CSRF validation",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful with tokens",
                        "schema": {
                            "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_identity_application_dto.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid state, missing code, or unsupported provider",
                        "schema": {
                            "$ref": "#/definitions/internal_identity_interface_http_v1.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "OAuth authentication failed",
                        "schema": {
                            "$ref": "#/definitions/internal_identity_interface_http_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_identity_interface_http_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_0xsj_hexagonal-go_internal_identity_application_dto.LoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "Token expiration in seconds",
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "token_type": {
                    "description": "\"Bearer\"",
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_0xsj_hexagonal-go_internal_identity_application_dto.UserDTO"
                }
            }
        },
...
        "github_com_0xsj_hexagonal-go_internal_identity_application_dto.UserDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "email_verified": {
                    "type": "boolean"
                },
                "email_verified_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_login_at": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_identity_interface_http_v1.ErrorResponse": {
            "description": "Error response returned by the API",
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "USER_NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "user not found"
                },
                "meta": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}